+++ project_task_department/__init__.py
@@ -1,1 +1,1 @@
-from . import project_task
+from . import models
