+++ project_task_department/__openerp__.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
+# © 2011 Joël Grand-Guillaume (Camptocamp)
+# © 2013 Daniel Reis (Camptocamp)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
     "name": "Project Task specific Department",
-    "version": "8.0.1.0.0",
+    "version": "9.0.1.0.0",
     "author": "Daniel Reis,Odoo Community Association (OCA)",
     "license": "AGPL-3",
     "category": "Generic Modules/Projects & Services",
@@ -17,6 +20,6 @@ available on them.)
     "depends": [
         "project_department",
     ],
-    "data": ["project_view.xml"],
-    'installable': False,
+    "data": ["views/project.xml"],
+    'installable': True,
 }
