+++ project_task_department/models/__init__.py
@@ -1,0 +1,1 @@
+from . import project_task
